+++ src/main/kotlin/io/saagie/astonparking/AstonParkingApplication.kt
@@ -1,90 +1,11 @@
 package io.saagie.astonparking
 
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.boot.SpringApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.autoconfigure.security.oauth2.resource.ResourceServerProperties
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.boot.web.servlet.FilterRegistrationBean
-import org.springframework.context.annotation.Bean
-import org.springframework.security.config.annotation.web.builders.HttpSecurity
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter
-import org.springframework.security.oauth2.client.OAuth2ClientContext
-import org.springframework.security.oauth2.client.OAuth2RestTemplate
-import org.springframework.security.oauth2.client.filter.OAuth2ClientAuthenticationProcessingFilter
-import org.springframework.security.oauth2.client.filter.OAuth2ClientContextFilter
-import org.springframework.security.oauth2.client.token.grant.code.AuthorizationCodeResourceDetails
-import org.springframework.security.oauth2.config.annotation.web.configuration.EnableOAuth2Client
-import org.springframework.security.web.authentication.LoginUrlAuthenticationEntryPoint
-import org.springframework.security.web.authentication.www.BasicAuthenticationFilter
-import org.springframework.security.web.csrf.CookieCsrfTokenRepository
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RestController
-import java.security.Principal
-import javax.servlet.Filter
 
 
 @SpringBootApplication
-@EnableOAuth2Client
-@RestController
-class AstonParkingApplication : WebSecurityConfigurerAdapter(false) {
-
-    @Qualifier("oauth2ClientContext")
-    @Autowired
-    var oauth2ClientContext: OAuth2ClientContext? = null
-
-    @GetMapping("/user")
-    fun user(principal: Principal): Principal {
-        return principal
-    }
-
-    @Throws(Exception::class)
-    override fun configure(http: HttpSecurity) {
-        // @formatter:off
-        http.antMatcher("/**").authorizeRequests().antMatchers("/", "/login**", "/webjars/**").permitAll().anyRequest()
-                .authenticated().and().exceptionHandling()
-                .authenticationEntryPoint(LoginUrlAuthenticationEntryPoint("/")).and().logout()
-                .logoutSuccessUrl("/").permitAll().and().csrf()
-                .csrfTokenRepository(CookieCsrfTokenRepository.withHttpOnlyFalse()).and()
-                .addFilterBefore(ssoFilter(), BasicAuthenticationFilter::class.java)
-        // @formatter:on
-    }
-
-    @Bean
-    fun oauth2ClientFilterRegistration(filter: OAuth2ClientContextFilter): FilterRegistrationBean {
-        val registration = FilterRegistrationBean()
-        registration.filter = filter
-        registration.order = -100
-        return registration
-    }
-
-    private fun ssoFilter(): Filter {
-        val slackFilter = OAuth2ClientAuthenticationProcessingFilter(
-                "/login/slack")
-        val slackTemplate = OAuth2RestTemplate(slack(), oauth2ClientContext)
-        slackFilter.setRestTemplate(slackTemplate)
-        val tokenServices = UserInfoTokenServices(slackResource().userInfoUri,
-                slack().clientId)
-        tokenServices.setRestTemplate(slackTemplate)
-        slackFilter.setTokenServices(
-                UserInfoTokenServices(slackResource().userInfoUri, slack().clientId))
-        return slackFilter
-    }
-
-    @Bean
-    @ConfigurationProperties("slack.client")
-    fun slack(): AuthorizationCodeResourceDetails {
-        return AuthorizationCodeResourceDetails()
-    }
-
-    @Bean
-    @ConfigurationProperties("slack.resource")
-    fun slackResource(): ResourceServerProperties {
-        return ResourceServerProperties()
-    }
-
-}
+class AstonParkingApplication
 
 fun main(args: Array<String>) {
     SpringApplication.run(AstonParkingApplication::class.java, *args)
