+++ src/main/resources/application.yaml
@@ -1,7 +1,7 @@
 slack:
   client:
-    clientId: "225187789888.225193189936"
-    clientSecret: "5104fde84aaf467439cdd820dea14c7e"
+    clientId: "3490096050.217656305830"
+    clientSecret: "46493610859295f7621f62a4be60d860"
     accessTokenUri: https://slack.com/api/oauth.access
     userAuthorizationUri: https://slack.com/oauth/authorize
     tokenName: token
@@ -12,4 +12,4 @@ slack:
     userInfoUri: https://slack.com/api/users.identity?token=
 logging:
   level:
-    org.springframework.security: DEBUG
+    org.springframework.security: WARN
